Add common form for editing and creating goal check-ins
