+++ assets/js/features/goals/GoalCheckIn/index.tsx
@@ -3,3 +3,4 @@ export { StatusSection } from "./StatusSection";
 export { DescriptionSection } from "./DescriptionSection";
 export { TargetsSection } from "./TargetsSection";
 export { useForm } from "./useForm";
+export { Form } from "./Form";
